Move activation handler into pkg/activator
