+++ pkg/activator/handler_test.go
@@ -10,7 +10,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package main
+package activator
 
 import (
 	"errors"
@@ -24,38 +24,37 @@ import (
 	"testing"
 
 	"github.com/knative/serving/cmd/util"
-	"github.com/knative/serving/pkg/activator"
 	"github.com/knative/serving/pkg/controller"
 	"go.uber.org/zap"
 )
 
-type fakeActivator struct {
-	endpoint  activator.Endpoint
+type stubActivator struct {
+	endpoint  Endpoint
 	namespace string
 	name      string
 }
 
-func newFakeActivator(namespace string, name string, server *httptest.Server) activator.Activator {
+func newStubActivator(namespace string, name string, server *httptest.Server) Activator {
 	url, _ := url.Parse(server.URL)
 	host := url.Hostname()
 	port, _ := strconv.Atoi(url.Port())
 
-	return &fakeActivator{
-		endpoint:  activator.Endpoint{FQDN: host, Port: int32(port)},
+	return &stubActivator{
+		endpoint:  Endpoint{FQDN: host, Port: int32(port)},
 		namespace: namespace,
 		name:      name,
 	}
 }
 
-func (fa *fakeActivator) ActiveEndpoint(namespace, name string) (activator.Endpoint, activator.Status, error) {
+func (fa *stubActivator) ActiveEndpoint(namespace, name string) (Endpoint, Status, error) {
 	if namespace == fa.namespace && name == fa.name {
 		return fa.endpoint, http.StatusOK, nil
 	}
 
-	return activator.Endpoint{}, http.StatusNotFound, errors.New("not found!")
+	return Endpoint{}, http.StatusNotFound, errors.New("not found!")
 }
 
-func (fa *fakeActivator) Shutdown() {
+func (fa *stubActivator) Shutdown() {
 }
 
 func TestActivationHandler(t *testing.T) {
@@ -72,7 +71,7 @@ func TestActivationHandler(t *testing.T) {
 	)
 	defer server.Close()
 
-	act := newFakeActivator("real-namespace", "real-name", server)
+	act := newStubActivator("real-namespace", "real-name", server)
 
 	examples := []struct {
 		label     string
